+++ CHANGELOG.md
@@ -67,6 +67,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   could cause a panic, potentially leading to denial of service.
   CVE ID: CVE-2021-43565
 
+### Added
+
+- Tarantool benchmark tool update (select and update operations)
+  * Flag "insert" - percentage of insert operations to bench space.
+  * Flag "select" - percentage of select operations from bench space.
+  * Flag "update" - percentage of update operations in bench space.
+  * Flag "fill" - number of records to pre-fill the space.
+
 ## [2.11.0] - 2022-01-26
 
 ### Changed